Refactor sponsors section
