+++ src/components/Sponsor/Container.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const Container = styled.section`
+  margin-bottom: 12rem;
+`
+
+export default Container
